+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/mission/dto/MissionDetailResponseDto.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.adapter.in.mission.dto;
 
-import com.booquest.booquest_api.adapter.in.mission.missionstep.dto.MissionStepResponseDto;
+import com.booquest.booquest_api.adapter.in.missionstep.dto.MissionStepResponseDto;
 import com.booquest.booquest_api.domain.mission.model.Mission;
 import com.fasterxml.jackson.databind.JsonNode;
