+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/missionstep/dto/MissionStepGenerateRequestDto.java
@@ -1,4 +1,4 @@
-package com.booquest.booquest_api.adapter.in.mission.missionstep.dto;
+package com.booquest.booquest_api.adapter.in.missionstep.dto;
 
 
 public record MissionStepGenerateRequestDto(
