+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/onboarding/web/dto/SideJobDetailResponseDto.java
@@ -2,7 +2,7 @@ package com.booquest.booquest_api.adapter.in.onboarding.web.dto;
 
 import com.booquest.booquest_api.adapter.in.mission.dto.MissionDetailResponseDto;
 import com.booquest.booquest_api.adapter.in.mission.dto.MissionResponseDto;
-import com.booquest.booquest_api.adapter.in.mission.missionstep.dto.MissionStepResponseDto;
+import com.booquest.booquest_api.adapter.in.missionstep.dto.MissionStepResponseDto;
 import com.booquest.booquest_api.domain.sidejob.model.SideJob;
 
 import java.util.List;
